first deployment test

+++ mistos-frontend/src/app/images/images-list/images-list.component.ts
@@ -1,10 +1,12 @@
 import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Data, Router } from '@angular/router';
 import { MatTableFilter } from 'mat-table-filter';
 import { Subscription } from 'rxjs';
+import { WarningDeleteComponent } from 'src/app/dialogs/warning-delete/warning-delete.component';
 import { Image } from 'src/app/models/image.model';
 import { ComService } from 'src/app/shared/com.service';
 import { ImageService } from 'src/app/shared/image.service';
@@ -24,13 +26,13 @@ export class ImagesListComponent implements AfterViewInit, OnInit {
   imageList: Image[];
   subscription: Subscription;
   displayedColumns = ['uid', 'name', "tags", "actions"];
-
+  dialogConfig = new MatDialogConfig();
 
   constructor(
     private comService: ComService,
-    private imageService: ImageService,
     private route: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private dialog: MatDialog
   ) {
 
   }
@@ -44,6 +46,12 @@ export class ImagesListComponent implements AfterViewInit, OnInit {
     this.filterType = MatTableFilter.ANYWHERE;
     //create dataSource for table
     this.dataSource = new MatTableDataSource(this.imageList);
+
+    //Setup Dialog Config File
+    this.dialogConfig.disableClose = true; //disables closing by clicking outside of the dialog
+    this.dialogConfig.autoFocus = true; //focus will automatically set on the first form field
+    this.dialogConfig.hasBackdrop = true; //prevents user from clicking on the rest of the ui
+    this.dialogConfig.closeOnNavigation = true; //closes dialog if wen navigate to another route
   }
 
   ngAfterViewInit() {
@@ -59,11 +67,6 @@ export class ImagesListComponent implements AfterViewInit, OnInit {
     this.dataSource.sort = this.sort;
   }
   
-  onPrint() {
-    console.log(this.imageList);
-    console.log(this.dataSource);
-  }
-
   onView(imageId) {
     this.comService.viewImage(imageId, true, true);
   }
@@ -73,15 +76,30 @@ export class ImagesListComponent implements AfterViewInit, OnInit {
   }
 
   onDelete(imageId: number) {
-    this.comService.deleteImageById(imageId).subscribe((response) => {
-      console.log("DeleteImageRequest:");
-      console.log(response);
-      
-      this.comService.fetchImageList().subscribe((imageList:Image[]) => {
-        this.imageList = imageList;
-        this.onFetchData();
+    this.dialogConfig.data = {
+      warningInput: "this image"
+    };
+
+    const dialogRef = this.dialog.open(         //dialogRef is a observable of the dialog
+      WarningDeleteComponent,
+      this.dialogConfig
+    );
+    dialogRef.afterClosed().subscribe((proceed:boolean)=> {
+      if (proceed === true) {
+        this.comService.deleteImageById(imageId).subscribe((response) => {
+          console.log("DeleteImageRequest:");
+          console.log(response);
+          
+          this.comService.fetchImageList().subscribe((imageList:Image[]) => {
+            this.imageList = imageList;
+            this.onFetchData();
+          });
       });
-  });
+      } else {
+        console.log("Delete Image was aboirted.");
+      }
+    }
+    ) 
   }
 }
